+++ plugwise/__init__.py
@@ -1,6 +1,6 @@
 """Plugwise module."""
 
-__version__ = "0.16.1a25"
+__version__ = "0.16.1a26"
 
 from plugwise.smile import Smile
 from plugwise.stick import Stick